refactor: implement LinkedIn login functionality and update user store structure
